feat: add delete prescription

+++ src/controllers/prescription.controller.ts
@@ -113,12 +113,9 @@ class PrescriptionController implements BaseController{
   public dispense = async (req: Request, res: Response) => {
     try{
       const prescriptionId: string = req.params.prescriptionId;
-      console.log("Prescription ID:", prescriptionId);
       const userId: string = req.params.userId;
-      console.log("UserId: ", userId);
       const status = 'Dispensada';
       const dispensedBy = await User.findOne({_id: userId});
-      console.log("DispensedBy: ", dispensedBy );
       const prescription = await Prescription.findOne({_id: prescriptionId});
       if(prescription?.status === 'Dispensada'){
         return res.status(422).json('La receta ya había sido dispensada.')
@@ -163,10 +160,14 @@ class PrescriptionController implements BaseController{
 
   public delete =  async (req: Request, res: Response): Promise<Response> => {
     try{
-
       const { id } = req.params;
-      await Prescription.findByIdAndDelete(id);
-      return res.status(200).json('deleted');
+      const prescription = await Prescription.findOne({_id: id});
+      if( prescription?.status === "Pendiente"){
+        await Prescription.findByIdAndDelete(id);
+        return res.status(200).json(prescription);
+      }else{
+        return res.status(422).json('La receta ya se ha dispensado y no puede ser eliminada.')
+      }
     }catch(err){
       console.log(err);
       return res.status(500).json('Server Error');
